pyproject.toml
[project]
name = "object_detection_module"
version = "1.5.0"
description = "A module that allows you to take a picture via webcam and run YOLO object detection through darknet"
authors = [{name = "Pranav Sinha", email="pranavsinha@oakland.edu"}]
dependencies = [
    "fastapi>=0.103",
    "uvicorn>=0.14.0",
    "opencv-python-headless",
    "ad_sdl.wei",
    "pytest",
    "numpy<2"
]
requires-python = ">=3.8.1"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/AD-SDL/python_template_module"

#####################
# Development Tools #
#####################

[tool.ruff]
# https://docs.astral.sh/ruff/configuration/

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "docs",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.8
target-version = "py38"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    # "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    # "SIM",
    # isort
    "I",
    # Warning
    "W",
    # pydocstyle
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107",
    # ruff
    # "RUF"
]
ignore = [
    "E501" # Line too long
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pytest.ini_options]
# https://docs.pytest.org/en/stable/customize.html
addopts = "-x"
junit_family="xunit1"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pottery.exceptions.InefficientAccessWarning",
]

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml
show_error_codes = true
check_untyped_defs = true
follow_imports = "normal"
strict_optional = true
plugins = ["pydantic.mypy"]
strict = true
disallow_untyped_defs = true
implicit_reexport = true
